requirements.txt
Django==4.2
djangorestframework>=3.12.4, <3.13
psycopg2>=2.8.6,<2.9
django-viewflow
# drf-spectacular>=0.26.3, <0.27
# Pillow>=8.2.0,<8.3.0
# uwsgi>=2.0.19, <2.1

